+++ src/gui/reports/ReportsDialog.h
@@ -31,6 +31,7 @@ class Entry;
 class Group;
 class QTabWidget;
 class ReportsPageHealthcheck;
+class ReportsPageHibp;
 class ReportsPageStatistics;
 
 namespace Ui
@@ -68,15 +69,17 @@ signals:
 
 private slots:
     void reject();
-    void entryActivationSignalReceived(const Group*, Entry* entry);
+    void entryActivationSignalReceived(Entry* entry);
     void switchToMainView(bool previousDialogAccepted);
 
 private:
     QSharedPointer<Database> m_db;
     const QScopedPointer<Ui::ReportsDialog> m_ui;
     const QSharedPointer<ReportsPageHealthcheck> m_healthPage;
+    const QSharedPointer<ReportsPageHibp> m_hibpPage;
     const QSharedPointer<ReportsPageStatistics> m_statPage;
     QPointer<EditEntryWidget> m_editEntryWidget;
+    QWidget* m_sender = nullptr;
 
     class ExtraPage;
     QList<ExtraPage> m_extraPages;